This commit adds the ability to sort tasks by order of priority, but also marks a big change to the underlying data structure of the TaskListModel class.
Previously we used a Map<String,TaskModel> to store TaskModels, but throughtout the proccess of designing the app I discovered that working with the map class
was cumbersome; I was always using the toList() and supporting different types of was going to be difficult SplayTreeSet and SplayTreeMap didn't help because I wanted something
easily indexable so the performance benefits didn't really matter at the end. So all in all I think a List<TaskModel> is going to be sufficient.